split main.ts
